move factory to http-kernel factory to support
